Refactore folders structure
