+++ src/bolireplate/helpers/configuration/configure-database-pg.helper.ts
@@ -5,7 +5,7 @@
  * @description  Configure PostgreSQL database connection
  */
 
-import { TypeOrmLoggerContainer } from '@/utils/type-orm-logger';
+import { TypeOrmLoggerContainer } from '@/bolireplate/utils/type-orm-logger';
 import { ConfigService } from '@nestjs/config';
 import { TypeOrmModuleOptions } from '@nestjs/typeorm';
