+++ src/bolireplate/types/enums/log-type.enum.ts
